feat: cleaned up old error handling
